LayoutTests/inspector/styles/styles-edit-property-after-invalid-rule-expected.txt

Tests that proper source lines are reported for the styles after unrecognizer / invalid selector.

Initial value
[expanded] 
element.style  { ()

======== Matched CSS Rules ========
[expanded] 
media="screen" (styles-edit-property-after-invalid-rule.html)
#container  { (styles-edit-property-after-invalid-rule.html:10)
padding: 15px;

[expanded] 
div  { (user agent stylesheet)
display: block;


After changing property
[expanded] 
element.style  { ()

======== Matched CSS Rules ========
[expanded] 
media="screen" (styles-edit-property-after-invalid-rule.html)
#container  { (styles-edit-property-after-invalid-rule.html:10)
padding: 20px;

[expanded] 
div  { (user agent stylesheet)
display: block;
